requirements.txt
nats-py
pymodbus==2.5.2
asyncio
aioredis==1.3.1
paho-mqtt